+++ src/types/interface.d.ts
@@ -1,4 +1,1 @@
 declare function parseInt(string: string | number, radix?: number | undefined): number;
-declare type Optional<T> = {
-    [k in keyof T]?: T[k];
-};